+++ CHANGELOG.md
@@ -1,5 +1,5 @@
 ## CHANGES in 0.10:
-
+- private messages are now included in downloads if the user has access to them (#169, #108)
 
 ## CHANGES in 0.9b: